rename and extend tests for NPE and NLE builder integration
